Usage
λ python tracker.py
Copy and paste the url of the game you want to track from cricbuzz manually ....

Enter the number of overs till which you want to track the score ...<eg. 16.4 or 20.0>...

You will get the live update of the score and at the end a graph will be plotted showing the variation in runs with successive overs.